Move types test to packages
