--- __tests__/types/__typescript__/fail-once/create-direct-line-invalid-bot-agent.tsx
@@ -1,4 +1,0 @@
-import { createDirectLine } from '../../../../packages/bundle';
-
-// "botAgent" is a forbidden option.
-createDirectLine({ botAgent: '123' });
