--- __tests__/types/__typescript__/fail-once/invalid-prop-type-dir.tsx
@@ -1,4 +1,0 @@
-import ReactWebChat from '../../../../packages/bundle';
-
-// "dir" must be a string of "ltr' | 'rtl' | 'auto'.
-<ReactWebChat dir={123} directLine={undefined} />;
