--- __tests__/types/__typescript__/fail-once/no-children-for-react-web-chat-in-minimal.tsx
@@ -1,6 +1,0 @@
-import ReactWebChat from '../../../../packages/bundle/lib/index-minimal';
-
-// <ReactWebChat> should not contains any children.
-<ReactWebChat directLine={undefined}>
-  <div>Hello, World</div>
-</ReactWebChat>;
