--- __tests__/types/__typescript__/fail-once/no-full-bundle-style-options-in-minimal-for-hooks.ts
@@ -1,5 +1,0 @@
-import { hooks } from '../../../../packages/bundle/lib/index-minimal';
-
-const [styleOptions] = hooks.useStyleOptions();
-
-styleOptions.cardEmphasisBackgroundColor = 'black';
