--- __tests__/types/__typescript__/pass/correct-type-dir.tsx
@@ -1,3 +1,0 @@
-import ReactWebChat from '../../../../packages/bundle';
-
-<ReactWebChat dir="ltr" directLine={undefined} />;
