--- __tests__/types/__typescript__/pass/import-component-send-text-box.tsx
@@ -1,5 +1,0 @@
-import { Components } from '../../../../packages/bundle';
-
-const { SendTextBox } = Components;
-
-<SendTextBox />;
