--- __tests__/types/__typescript__/pass/import-create-direct-line.tsx
@@ -1,3 +1,0 @@
-import { createDirectLine } from '../../../../packages/bundle';
-
-createDirectLine({ token: 'faketoken' });
