--- __tests__/types/__typescript__/pass/render-with-store.tsx
@@ -1,9 +1,0 @@
-import ReactDOM from 'react-dom';
-
-import { Components, createStore } from '../../../../packages/bundle';
-
-const { Composer } = Components;
-
-const store = createStore({});
-
-ReactDOM.render(<Composer directLine={undefined} store={store} />, document.getElementById('app'));
