--- __tests__/types/__typescript__/pass/render.tsx
@@ -1,7 +1,0 @@
-import ReactDOM from 'react-dom';
-
-import ReactWebChat, { createDirectLine } from '../../../../packages/bundle';
-
-const directLine = createDirectLine({ token: '...' });
-
-ReactDOM.render(<ReactWebChat directLine={directLine} />, document.getElementById('app'));
