--- __tests__/types/__typescript__/pass/valid-style-options-suggested-actions-stacked-overflow.ts
@@ -1,7 +1,0 @@
-import { createStyleSet } from '../../../../packages/bundle';
-
-// Related to #4081.
-createStyleSet({ suggestedActionsStackedOverflow: 'auto' });
-createStyleSet({ suggestedActionsStackedOverflow: 'hidden' });
-createStyleSet({ suggestedActionsStackedOverflow: 'scroll' });
-createStyleSet({ suggestedActionsStackedOverflow: 'visible' });
