+++ packages/bundle/__tests__/types/__typescript__/fail-once/create-direct-line-invalid-bot-agent.tsx
@@ -1,0 +1,4 @@
+import { createDirectLine } from '../../../../lib/index';
+
+// "botAgent" is a forbidden option.
+createDirectLine({ botAgent: '123' });
