+++ packages/bundle/__tests__/types/__typescript__/fail-once/create-direct-line-invalid-option.tsx
@@ -1,0 +1,4 @@
+import { createDirectLine } from '../../../../lib/index';
+
+// Object literal may only specify known properties, and 'invalid' does not exist in type 'Omit<CreateDirectLineOptions, "botAgent">'.
+createDirectLine({ invalid: true });
