+++ packages/bundle/__tests__/types/__typescript__/fail-once/no-children-for-react-web-chat.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import ReactWebChat from '../../../../lib/index';
+
+// <ReactWebChat> should not contains any children.
+<ReactWebChat directLine={undefined}>
+  <div>Hello, World</div>
+</ReactWebChat>;
