+++ packages/bundle/__tests__/types/__typescript__/fail-once/no-full-bundle-style-options-in-minimal-for-hooks.ts
@@ -1,0 +1,5 @@
+import { hooks } from '../../../../lib/index-minimal';
+
+const [styleOptions] = hooks.useStyleOptions();
+
+styleOptions.cardEmphasisBackgroundColor = 'black';
