+++ packages/bundle/__tests__/types/__typescript__/fail-once/no-full-bundle-style-options-in-minimal.ts
@@ -1,0 +1,4 @@
+import { createStyleSet } from '../../../../lib/index-minimal';
+
+// "cardEmphasisBackgroundColor" is a style options only available in full bundle.
+createStyleSet({ cardEmphasisBackgroundColor: 'orange' });
