+++ packages/bundle/__tests__/types/__typescript__/pass/correct-type-dir.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import ReactWebChat from '../../../../lib/index';
+
+<ReactWebChat dir="ltr" directLine={undefined} />;
