+++ packages/bundle/__tests__/types/__typescript__/pass/use-style-options-with-full-bundle.ts
@@ -1,0 +1,5 @@
+import { hooks } from '../../../../lib/index';
+
+const [styleOptions] = hooks.useStyleOptions();
+
+styleOptions.cardEmphasisBackgroundColor = 'black';
