+++ packages/bundle/__tests__/types/__typescript__/pass/valid-style-options-suggested-actions-stacked-overflow.ts
@@ -1,0 +1,7 @@
+import { createStyleSet } from '../../../../lib/index';
+
+// Related to #4081.
+createStyleSet({ suggestedActionsStackedOverflow: 'auto' });
+createStyleSet({ suggestedActionsStackedOverflow: 'hidden' });
+createStyleSet({ suggestedActionsStackedOverflow: 'scroll' });
+createStyleSet({ suggestedActionsStackedOverflow: 'visible' });
